Transition to subcommand CLI
Move argparse configuration into new args.py module. Defines config,
init-series, and fetch-subs subcommands.

config provides interactive configuration of TMDB & OpenSubtitles auth
params, which are overrideable with command line arguments.

init-series initializes a directory as containing a specified TV series,
which is inferred from the directory name by default. All series-level
data is stored in a .mkv-episode-matcher directory. This replaces the
~/.mkv-episode-matcher/cache directory structure for storing subtitles
files. The primary purpose is to download TMDB series & season data for
use in other operations. The show id can be specified as a command line
argument, or there is an interactive UI for choosing the correct series
when there are multiple matches for the series name.

fetch-subs downloads subtitles from OpenSubtitles. These are stored
alongside the OpenSubtitles response JSON for future references. The
subs to download can be limited by the --seasons and --episodes command
line arguments.